app/__init__.py
from flask import Flask
from dotenv import load_dotenv
import os
from flask_sqlalchemy import SQLAlchemy
from flask_login import LoginManager

load_dotenv()

app = Flask(__name__)
app.config['SECRET_KEY'] = os.environ.get('SECRET_KEY')
app.config['SQLALCHEMY_DATABASE_URI'] = os.environ.get('MYSQL_CONN_STRING')

db = SQLAlchemy(app)
print(db)

login = LoginManager(app)
login.login_view = 'login'

from app import routes, models